requirements.txt.txt
python-telegram-bot==20.3
ffmpeg-python==0.2.0